remove ugly screenshots and error handling in GetData

+++ CourseWorkTerm4/sort.h
@@ -9,12 +9,12 @@ typedef enum
 } sort_method;
 
 // Функции для сортировки
-bool AscendingOrder(int a, int b);
-bool DescendingOrder(int a, int b);
+bool AscendingOrder(double a, double b);
+bool DescendingOrder(double a, double b);
 
 // Функция получения данных из input
-int GetData(std::vector<int>& array, std::wstring str);
+int GetData(std::vector<double>& array, std::wstring str);
 
 // Функции для вызова sort с разными функциями
-int SortWithRegular(std::vector<int>& array, sort_method sm);
-int SortWithLambda(std::vector<int>& array, sort_method sm);
+int SortWithRegular(std::vector<double>& array, sort_method sm);
+int SortWithLambda(std::vector<double>& array, sort_method sm);
